:truck: move modules again
Try to keep the most similar structure of modules import, like old version
