+++ LICENSE
@@ -3,7 +3,7 @@ GNU GENERAL PUBLIC LICENSE
 
     Python USI submission REST API contain all methods to interact with EMBL-EBI
     Unified Submissions Interface
-    Copyright (C) 2018-2019  Paolo Cozzi
+    Copyright (C) 2018-2020  Paolo Cozzi
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
