+++ docs/index.rst
@@ -2,7 +2,7 @@ Welcome to Python USI submission REST API's documentation!
 ==========================================================
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 3
    :caption: Contents:
 
    readme
