--- pyUSIrest/client/__init__.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jan  7 14:46:41 2020
-
-@author: Paolo Cozzi <paolo.cozzi@ibba.cnr.it>
-"""
-
-from .client import Client
-from .document import Document
-
-__all__ = ["Client", "Document"]
